Fix 32-bit issue

+++ src/graphql/graphqlModules/Miner/MinerStats.js
@@ -86,30 +86,30 @@ module.exports.typeDefs = `
     interval: Int
     jobs: Int
     cleanFlags: Int
-    sharesSent: Int
-    sharesAccepted: Int
-    sharesRejected: Int
-    solutionsAccepted: Int
+    sharesSent: Float
+    sharesAccepted: Float
+    sharesRejected: Float
+    solutionsAccepted: Float
     minRespTime: Float
     avgRespTime: Float
     maxRespTime: Float
     shareLoss: Float
-    poolTotal: Int
+    poolTotal: Float
     inService: Int
     subscribeError: Int
-    diffChanges: Int
+    diffChanges: Float
     reconnections: Int
     reconnectionsOnErrors: Int
-    defaultJobShares: Int
-    staleJobShares: Int
-    duplicateShares: Int
-    lowDifficultyShares: Int
-    pwcSharesSent: Int
-    pwcSharesDropped: Int
-    bigDiffShares: Int
-    belowTargetShare: Int
+    defaultJobShares: Float
+    staleJobShares: Float
+    duplicateShares: Float
+    lowDifficultyShares: Float
+    pwcSharesSent: Float
+    pwcSharesDropped: Float
+    bigDiffShares: Float
+    belowTargetShare: Float
     pwcRestart: Int
-    statOverflow: Int
+    statOverflow: Float
   }
 
   type MinerStatsFans {
@@ -170,7 +170,7 @@ module.exports.typeDefs = `
   type MinerStatsSlotAlert {
     alertLo: Int
     alertHi: Int
-    numWrite: Int
+    numWrite: Float
   }
 
   type MinerStatsSlave {
